Added UserService, ProductService and SupplierService

+++ src/main/java/com/quest/tims/controller/UserController.java
@@ -3,7 +3,6 @@ package com.quest.tims.controller;
 import com.quest.tims.entity.User;
 import com.quest.tims.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,27 +15,27 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-    // GET /users: Retrieve all users (Admin only)
+    // GET /users: Retrieve all users
     @GetMapping
     public ResponseEntity<List<User>> getAllUsers() {
         return userService.getAllUsers();
     }
 
-    // GET /users/{id}: Retrieve a single user by ID (Admin only)
+    // GET /users/{id}: Retrieve a single user by ID
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable Long id) {
+    public ResponseEntity<User> getUserById(@PathVariable int id) {
         return userService.getUserById(id);
     }
 
-    // POST /users: Add a new user (Admin only)
+    // POST /users: Add a new user
     @PostMapping
     public ResponseEntity<User> createUser(@RequestBody User user) {
         return userService.createUser(user);
     }
 
-    // DELETE /users/{id}: Delete a user by ID (Admin only)
+    // DELETE /users/{id}: Delete a user by ID
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<String> deleteUser(@PathVariable int id) {
         return userService.deleteUser(id);
     }
 }
